Fix worldpop service request plus url database syntax

+++ osm/src/main/groovy/org/orbisgis/geoclimate/osm/WorkflowOSM.groovy
@@ -539,33 +539,42 @@ Map osm_processing(JdbcDataSource h2gis_datasource, def processing_parameters, d
                 //Extract and compute population indicators for the specified year
                 //This data can be used by the grid_indicators process
                 if (worldpop_indicators) {
-                    def bbox = [zones.osm_envelope_extented.getMinY() as Float, zones.osm_envelope_extented.getMinX() as Float,
-                                zones.osm_envelope_extented.getMaxY() as Float, zones.osm_envelope_extented.getMaxX() as Float]
                     String coverageId = "wpGlobal:ppp_2020"
-                    String worldPopFile = WorldPopTools.Extract.extractWorldPopLayer(coverageId, bbox)
-                    if (worldPopFile) {
-                        String worldPopTableName = WorldPopTools.Extract.importAscGrid(h2gis_datasource, worldPopFile, srid, coverageId.replaceAll(":", "_"))
-                        if (worldPopTableName) {
-                            results.put("population", worldPopTableName)
-                            String buildingWithPop = Geoindicators.BuildingIndicators.buildingPopulation(h2gis_datasource, results.building, worldPopTableName, ["pop"])
-                            h2gis_datasource.dropTable(worldPopTableName)
-                            if (!buildingWithPop) {
-                                info "Cannot compute any population data at building level"
+                    if(WorldPopTools.Extract.isCoverageAvailable(coverageId)) {
+                        def bbox = [zones.osm_envelope_extented.getMinY() as Float, zones.osm_envelope_extented.getMinX() as Float,
+                                    zones.osm_envelope_extented.getMaxY() as Float, zones.osm_envelope_extented.getMaxX() as Float]
+
+                        String worldPopFile = WorldPopTools.Extract.extractWorldPopLayer(coverageId, bbox)
+                        if (worldPopFile) {
+                            String worldPopTableName = WorldPopTools.Extract.importAscGrid(h2gis_datasource, worldPopFile, srid, coverageId.replaceAll(":", "_"))
+                            if (worldPopTableName) {
+                                results.put("population", worldPopTableName)
+                                String buildingWithPop = Geoindicators.BuildingIndicators.buildingPopulation(h2gis_datasource, results.building, worldPopTableName, ["pop"])
+                                h2gis_datasource.dropTable(worldPopTableName)
+                                if (!buildingWithPop) {
+                                    info "Cannot compute any population data at building level"
+                                } else {
+                                    h2gis_datasource.dropTable(results.building)
+                                    //Update the building table with the population data
+                                    results.put("building", buildingWithPop)
+                                }
+
                             } else {
-                                h2gis_datasource.dropTable(results.building)
-                                //Update the building table with the population data
-                                results.put("building", buildingWithPop)
+                                info "Cannot import the worldpop asc file $worldPopFile"
+                                info "Create a default empty worldpop table"
+                                def outputTableWorldPopName = postfix "world_pop"
+                                h2gis_datasource.execute("""drop table if exists $outputTableWorldPopName;
+                                        create table $outputTableWorldPopName (the_geom GEOMETRY(POLYGON, $srid), ID_POP INTEGER, POP FLOAT);""".toString())
+                                results.put("population", outputTableWorldPopName)
                             }
 
                         } else {
-                            info "Cannot import the worldpop asc file $worldPopFile"
-                            info "Create a default empty worldpop table"
+                            info "Cannot find the population grid $coverageId \n Create a default empty worldpop table"
                             def outputTableWorldPopName = postfix "world_pop"
                             h2gis_datasource.execute("""drop table if exists $outputTableWorldPopName;
-                                        create table $outputTableWorldPopName (the_geom GEOMETRY(POLYGON, $srid), ID_POP INTEGER, POP FLOAT);""".toString())
+                                    create table $outputTableWorldPopName (the_geom GEOMETRY(POLYGON, $srid), ID_POP INTEGER, POP FLOAT);""".toString())
                             results.put("population", outputTableWorldPopName)
                         }
-
                     } else {
                         info "Cannot find the population grid $coverageId \n Create a default empty worldpop table"
                         def outputTableWorldPopName = postfix "world_pop"
